backend/requirements.txt
fastapi
uvicorn
openai-whisper
pydub
python-multipart
requests
openai
websockets
webrtcvad
pyannote.audio
torch